requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.7
jieba3k==0.35.1
lxml==4.2.5
newspaper3k==0.2.8
nltk==3.3
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-file==1.4.3
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tinysegmenter==0.3
tldextract==2.2.0
urllib3==1.24.1
